backend/requirements.txt
fastapi
uvicorn
sqlmodel

python-multipart
python-dotenv
asyncpg
psycopg2-binary
requests
